fix last_hitting_attack_id

+++ src/lib.rs
@@ -89,7 +89,7 @@ pub struct Frame {
     pub hitlag_frames: f32,
     pub last_ground_idx: u16,
     pub state_flags: [u8; 5],
-    pub last_hitting_attack_id: u16,
+    pub last_hitting_attack_id: u8,
     pub last_hit_by_instance_id: u16,
 }
